+++ src/test/converter2/validation/unusedMergeModuleWith.ts
@@ -1,0 +1,5 @@
+/**
+ * @module
+ * @mergeModuleWith notUsed
+ */
+export const test = 1;