+++ src/test/utils/options/default-options.test.ts
@@ -131,6 +131,168 @@ describe("Default Options", () => {
         });
     });
 
+    describe("locales", () => {
+        it("Should disallow non-objects", () => {
+            throws(() => opts.setValue("locales", null as never));
+        });
+
+        it("Should disallow objects containing non-objects", () => {
+            throws(() => opts.setValue("locales", { test: false } as never));
+        });
+
+        it("Should disallow locales containing non-strings", () => {
+            throws(() => opts.setValue("locales", { test: { key: false } } as never));
+        });
+
+        it("Should allow valid locale shapes", () => {
+            doesNotThrow(() => opts.setValue("locales", { test: { key: "hi" } }));
+        });
+    });
+
+    describe("packageOptions", () => {
+        it("Should disallow non-objects", () => {
+            throws(() => opts.setValue("packageOptions", null as never));
+        });
+
+        it("Should allow objects", () => {
+            doesNotThrow(() => opts.setValue("packageOptions", { test: false } as never));
+        });
+    });
+
+    describe("blockTags", () => {
+        it("Should disallow non-tags", () => {
+            throws(() => opts.setValue("blockTags", ["@bad-non-tag"]));
+        });
+
+        it("Should allow tags", () => {
+            doesNotThrow(() => opts.setValue("blockTags", ["@good"]));
+        });
+    });
+
+    describe("excludeNotDocumentedKinds", () => {
+        it("Should disallow invalid kind strings", () => {
+            throws(() => opts.setValue("excludeNotDocumentedKinds", ["InvalidKind"] as never));
+        });
+
+        it("Should disallow disallowed kind strings", () => {
+            throws(() => opts.setValue("excludeNotDocumentedKinds", ["Project"] as never));
+        });
+
+        it("Should kinds", () => {
+            doesNotThrow(() => opts.setValue("excludeNotDocumentedKinds", ["Variable"]));
+        });
+    });
+
+    describe("externalSymbolLinkMappings", () => {
+        it("Should disallow non-objects", () => {
+            throws(() => opts.setValue("externalSymbolLinkMappings", null as never));
+        });
+
+        it("Should disallow objects containing non-objects", () => {
+            throws(() => opts.setValue("externalSymbolLinkMappings", { test: false } as never));
+        });
+
+        it("Should disallow mappings containing non-strings", () => {
+            throws(() => opts.setValue("externalSymbolLinkMappings", { test: { key: false } } as never));
+        });
+
+        it("Should allow valid mapping shapes", () => {
+            doesNotThrow(() => opts.setValue("externalSymbolLinkMappings", { test: { key: "hi" } }));
+        });
+    });
+
+    describe("outputs", () => {
+        it("Should disallow non-arrays", () => {
+            throws(() => opts.setValue("outputs", null as never));
+        });
+
+        it("Should arrays of invalid shape", () => {
+            throws(() => opts.setValue("outputs", [{}] as never));
+        });
+
+        it("Should allow valid shapes", () => {
+            doesNotThrow(() => opts.setValue("outputs", [{ name: "html", path: "out" }]));
+        });
+    });
+
+    describe("highlightLanguages", () => {
+        it("Should disallow non-arrays", () => {
+            throws(() => opts.setValue("highlightLanguages", null as never));
+        });
+
+        it("Should disallow arrays containing unknown languages", () => {
+            throws(() => opts.setValue("highlightLanguages", ["notASupportedLanguage"] as never));
+        });
+
+        it("Should allow valid languages", () => {
+            doesNotThrow(() => opts.setValue("highlightLanguages", ["typescript"]));
+        });
+    });
+
+    describe("markdownItLoader", () => {
+        it("Should disallow non-functions", () => {
+            throws(() => opts.setValue("markdownItLoader", null as never));
+        });
+
+        it("Should allow functions", () => {
+            doesNotThrow(() => opts.setValue("markdownItLoader", () => {}));
+        });
+    });
+
+    describe("favicon", () => {
+        it("Should disallow paths with an unknown extension", () => {
+            throws(() => opts.setValue("favicon", "test.txt"));
+        });
+
+        it("Should allow remote paths with an unknown extension", () => {
+            doesNotThrow(() => opts.setValue("favicon", "https://example.com/test.txt"));
+        });
+
+        it("Should allow paths with a known extension", () => {
+            doesNotThrow(() => opts.setValue("favicon", "test.png"));
+        });
+    });
+
+    describe("hostedBaseurl", () => {
+        it("Should disallow non-http urls", () => {
+            throws(() => opts.setValue("hostedBaseUrl", "test"));
+            throws(() => opts.setValue("hostedBaseUrl", "ftp://test"));
+        });
+
+        it("Should allow http and https urls", () => {
+            doesNotThrow(() => opts.setValue("hostedBaseUrl", "http://example.com/"));
+            doesNotThrow(() => opts.setValue("hostedBaseUrl", "https://example.com/"));
+        });
+    });
+
+    describe("visibilityFilters", () => {
+        it("Should disallow non-objects", () => {
+            throws(() => opts.setValue("visibilityFilters", "test" as never));
+        });
+
+        it("Should disallow objects with an invalid key", () => {
+            throws(() => opts.setValue("visibilityFilters", { bad: true } as never));
+        });
+
+        it("Should disallow objects with a non-boolean values", () => {
+            throws(() => opts.setValue("visibilityFilters", { private: "bad" } as never));
+        });
+
+        it("Should allow objects with a valid shape", () => {
+            doesNotThrow(() => opts.setValue("visibilityFilters", { private: true }));
+            doesNotThrow(() => opts.setValue("visibilityFilters", { "@test": true }));
+        });
+    });
+
+    describe("kindSortOrder", () => {
+        it("Should disallow non-kind strings", () => {
+            throws(() => opts.setValue("kindSortOrder", ["Bad"] as never));
+        });
+        it("Should allow valid sort orders", () => {
+            doesNotThrow(() => opts.setValue("kindSortOrder", ["Accessor"]));
+        });
+    });
+
     it("Package Options Documentation", () => {
         const allOptions = opts
             .getDeclarations()